ui: Split main-button components into different files
